Add functioning targeting system and animation system. Zombies can now be shot at and killed.
Pressing F brings up aim/fire mode. The crosshair appears and automatically locks onto the nearest zombie in sight. It can be manually moved using the cursor keys. Pressing F again fires at the target under the crosshair (at which point the turn ends). Tapping F repeatedly will rapidly target and fire at all nearest zombies around the player.

The animation system is geared towards displaying animations between turns to provide feedback to the player as to what just happened. Not important currently but later if a lot is happening on screen it will be useful to see what is shooting what. Two initial animations added are a muzzleflash animation shown when the gun is fired and an impact animation when a zombie is hit.

There is a bug that dead zombies block the player's movement.